Add optional `bom.ini` configuration of `boards` (qty of PCBs in the BOM) and `pcbConfig` (the PCB variant).
Bom.ini values are used if the command line arguments are *not* provided. This allows project defaults to be set as
generally the `pcbConfig` options will be project specific, but KiCAD persists the command line options between projects
requiring manual entry of these each time you switch project. This is also helpful for projects where version control is
used as the variant and number will persist in bom.ini if it is in git/whatever vcs.