Add support for UTF-8 encoding, for py2 and py3
Needs to check version, then handle accordingly. Crux of the issue is
that the `csv` module in python2 does not support UTF-8. See note here:
https://docs.python.org/2/library/csv.html

> **Note**: This version of the csv module doesn’t support Unicode
> input. Also, there are currently some issues regarding ASCII NUL
> characters. Accordingly, all input should be UTF-8 or printable
> ASCII to be safe; see the examples in section Examples.

Python3 `csv` does support UTF-8, and you can open the file with a
UTF-8 encoding---something which you can not do in python2.

This is a bit clunky, perhaps with python2 finally on the way out,
support could be dropped, and python3 asserted. In the mean time this
is a sufficient work-around.