Exclude components marked as "Exclude from BoM"
Note that this flag can't be seen on files generated using the GUI and
asking to create a BoM.
But if you ask to export the netlist using the XML format, i.e. using
kicad-cli, you'll get the components listed.
Also note that the same happens with the DNP flag!