Replacing field values using the complex_variant = True option is not case sensitive.
If the field name to be replaced contains uppercase letters, the comparison will fail because of the following code

if field.get('field', 'name') == name:

The first part of the comparison may contain uppercase letters. The latter part is typically lower case, but for safety reasons, I apply lower() as well. This PR contains the fix and also squashes some trailing white spaces.